Move SDK code to the root of the repo and fix Go module name
