--- src/salad-job-queue-worker/.gitignore
@@ -1,2 +1,0 @@
-salad-job-queue-worker
-
